Initial implementation to support n separate data output lines in a single SPI master.
These changes fail to build on my machine at the moment, throwing many warnings and using too many threads, timers and channels on tile 1. The failure may be due to my inability to run to a successful completion the Build.pl script for domains tools_released, tools_xmostest and xdoc_released. My inability to run Build.pl lies in the power surge of last weekend which has resulted in the need to restore all of /projects.